+++ Assets/Src/Data/BulletData.cs
@@ -10,7 +10,7 @@ namespace Data
         /// <summary>
         /// 子弹ID，可通过数据表中读取子弹的信息。
         /// </summary>
-        public string Id;
+        public int Id;
 
         /// <summary>
         /// 父英雄。