+++ CHANGELOG.md
@@ -4,6 +4,8 @@
     - 通过AssetsManager.Instance.LoadLuBan("Assets/AllConfigs.bytes")一次性加载所有LuBan数据资源；
     - 内部会自动解析LuBan数据，生成对应的表格数据结构；二进制结构为（名字|数据长度|数据）
 3、新增SpriteAtlas资源加载的支持；
+    - 新增GetSprite方法，用于获得SpriteAtlas中的Sprite资源；
+    - 新增GetSprites方法，用于获得SpriteAtlas中的多个Sprite资源；
 
 ## 26.4.20
 1、新增：为`FairyUIFormLogic`组件添加了一个自动绑定字段的方法，用于自动绑定FairyGUI的字段，比如布局里存在一个btn_start按钮，自动绑定到对应逻辑类的`btn_start`字段上；